Files I created
Here are the files and classes that I created. 
The DescrptionController.java class  that I created is to show the region and description of the animals when the user is shown the list of animals after they select a category to choose from.
The ZooDexModel.java class is a model class that I created is used to handle all the methods being used by the controller classes.
The .txt files are the text files that I created with the list of animals that I chose with the region and description. Feel free to add more if you would like.